requirements.txt
streamlit==1.32.0
pandas==2.2.1
pillow==10.2.0
python-dotenv==1.0.1
firebase-admin==6.4.0
python-jose==3.3.0
requests==2.31.0
gunicorn==21.2.0
werkzeug==2.3.7 